hotfix for meal seating
