+++ ap/meal_seating/models.py
@@ -1,11 +1,11 @@
 from django.db import models
-from accounts.models import User, Trainee
 
 """ meal_seating models.py
 
 This module prints the bi-weekly meal seating list for the FTTA and FTTMA
 brothers and sisters, and also for short-termers.
 """
+
 
 class Table(models.Model):
 
@@ -34,30 +34,6 @@ class Table(models.Model):
     except AttributeError as e:
       return str(self.id) + ": " + str(e)
 
-  @staticmethod
-  def seatinglist(genderlist, gender):
-    tables = Table.objects.filter(gender = gender)
-    traineenum = 0
-    tablenum = 0
-    totalcapacity = 0
-    meal_list = []
-    for x in Table.objects.all().filter(gender = gender).values("capacity"):
-      totalcapacity += x["capacity"]
-    if (len(genderlist) > totalcapacity):
-      print "cannot seat " , len(genderlist) , " trainees. Current capacity is: " , totalcapacity
-      return None
-    else:
-      for trainee in genderlist:
-        meal_seating = {}
-        if ( traineenum == tables[tablenum].capacity):
-          tablenum += 1
-          traineenum = 0
-        meal_seating["first_name"] = trainee.firstname
-        meal_seating["last_name"] = trainee.lastname
-        meal_seating["table"] = tables[tablenum]
-        meal_list.append(meal_seating)
-        traineenum += 1
-      return meal_list
 
 class TraineeExclusion(models.Model):
   trainee = models.IntegerField()
