+++ ap/meal_seating/views.py
@@ -5,15 +5,11 @@ from django.contrib import messages
 from meal_seating.models import Table, TraineeExclusion
 from accounts.models import Trainee
 from datetime import date, timedelta
-from django.views.generic import TemplateView, DetailView, ListView
 from django.views import generic
-from terms.models import Term
-from accounts.models import Trainee
 from .serializers import TableSerializer
 
-from rest_framework import viewsets, filters
+from rest_framework import viewsets
 from rest_framework.renderers import JSONRenderer
-from rest_framework_bulk import BulkModelViewSet
 
 
 @csrf_protect
@@ -30,9 +26,11 @@ def editinfo(request):
   TraineeExclusion.objects.bulk_create(trainee_exclusions)
   return HttpResponse(difference)
 
+
 def get_exclude_list():
   exclude_list = TraineeExclusion.objects.values_list('trainee', flat=True)
   return exclude_list if exclude_list else []
+
 
 def seattables(request):
   filterchoice = request.POST['Filter']
@@ -41,27 +39,55 @@ def seattables(request):
   enddate = request.POST['enddate']
 
   trainees = Trainee.objects.all().filter(gender=genderchoice).order_by(filterchoice).exclude(id__in=get_exclude_list())
-  seating_list = Table.seatinglist(trainees,genderchoice)
-  if seating_list == None:
+  seating_list = seatinglist(trainees, genderchoice)
+  if seating_list is None:
     messages.error(request, 'Not enough seats available!')
     return redirect('/meal_seating')
   else:
-    return render(request, 'meal_seating/detail.html', {'seating_list' : seating_list, 'startdate':startdate, 'enddate': enddate})
+    return render(request, 'meal_seating/detail.html', {'seating_list': seating_list, 'startdate': startdate, 'enddate': enddate})
+
 
 def newseats(request):
   exclude_list = get_exclude_list()
   trainees_exclude = Trainee.objects.filter(pk__in=exclude_list)
   trainees = Trainee.objects.filter(is_active=1).exclude(type="S").order_by("lastname")
-  return render(request, 'meal_seating/index.html', {'trainees' : trainees,'exclude_list':trainees_exclude,})
+  return render(request, 'meal_seating/index.html', {'trainees': trainees, 'exclude_list': trainees_exclude, })
+
 
 def signin(request):
   trainees = Trainee.objects.all().filter(is_active=1).order_by("lastname")
   startdate = date.today()
   two_week_datelist = []
-  for x in range(0,14):
+  for x in range(0, 14):
     mydate = startdate + timedelta(days=x)
     two_week_datelist.append(format(mydate))
-  return render(request, 'meal_seating/mealsignin.html', {'trainees' : trainees, 'start_date' : startdate, "two_week_datelist" : two_week_datelist})
+  return render(request, 'meal_seating/mealsignin.html', {'trainees': trainees, 'start_date': startdate, "two_week_datelist": two_week_datelist})
+
+
+def seatinglist(genderlist, gender):
+    tables = Table.objects.filter(gender=gender)
+    traineenum = 0
+    tablenum = 0
+    totalcapacity = 0
+    meal_list = []
+    for x in Table.objects.all().filter(gender=gender).values("capacity"):
+      totalcapacity += x["capacity"]
+    if (len(genderlist) > totalcapacity):
+      print "cannot seat ", len(genderlist), " trainees. Current capacity is: ", totalcapacity
+      return None
+    else:
+      for trainee in genderlist:
+        meal_seating = {}
+        if traineenum == tables[tablenum].capacity:
+          tablenum += 1
+          traineenum = 0
+        meal_seating["first_name"] = trainee.firstname
+        meal_seating["last_name"] = trainee.lastname
+        meal_seating["table"] = tables[tablenum]
+        meal_list.append(meal_seating)
+        traineenum += 1
+      return meal_list
+
 
 class TableListView(generic.ListView):
   model = Table
@@ -76,6 +102,7 @@ class TableListView(generic.ListView):
     context = super(TableListView, self).get_context_data(**kwargs)
     context['table_bb'] = l_render(TableSerializer(table, many=True).data)
     return context
+
 
 class TableViewSet(viewsets.ModelViewSet):
   """
